Removed db spacial class
