+++ lib/presentation/pages/choose_automation_type_to_add_page.dart
