+++ lib/presentation/pages/pages.dart
@@ -1,30 +1,30 @@
-export 'package:cybearjinni/presentation/pages/add_new_automation_process/add_action/add_action_page.dart';
-export 'package:cybearjinni/presentation/pages/add_new_automation_process/add_bindings/add_binding_page.dart';
-export 'package:cybearjinni/presentation/pages/add_new_automation_process/add_routine/add_routine_page.dart';
-export 'package:cybearjinni/presentation/pages/add_new_automation_process/add_routine/pick_repeat_time/pick_repeat_time_page.dart';
-export 'package:cybearjinni/presentation/pages/add_new_automation_process/add_scene/add_scene_page.dart';
-export 'package:cybearjinni/presentation/pages/add_new_automation_process/choose_automation_type_to_add/choose_automation_type_to_add_page.dart';
+export 'package:cybearjinni/presentation/pages/add_action_page.dart';
+export 'package:cybearjinni/presentation/pages/add_bindings/add_binding_page.dart';
 export 'package:cybearjinni/presentation/pages/add_new_devices_process/choose_device_vendor_to_add/choose_device_vendor_to_add_page.dart';
 export 'package:cybearjinni/presentation/pages/add_new_devices_process/computer_connection_check/computer_connection_check_page.dart';
 export 'package:cybearjinni/presentation/pages/add_new_devices_process/configure_new_cbj_comp/configure_new_cbj_comp_page.dart';
 export 'package:cybearjinni/presentation/pages/add_new_devices_process/connect_to_home_wifi/connect_to_home_wifi_page.dart';
 export 'package:cybearjinni/presentation/pages/add_new_devices_process/open_access_point/open_access_pointi_page.dart';
 export 'package:cybearjinni/presentation/pages/add_new_devices_process/scan_for_new_cbj_comps/scan_for_new_cbj_comps_page.dart';
-export 'package:cybearjinni/presentation/pages/add_new_room/add_new_room_page.dart';
+export 'package:cybearjinni/presentation/pages/add_new_room_page.dart';
+export 'package:cybearjinni/presentation/pages/add_routine/add_routine_page.dart';
+export 'package:cybearjinni/presentation/pages/add_scene/add_scene_page.dart';
 export 'package:cybearjinni/presentation/pages/add_vendors_pages/add_esphome_vendor/add_esphome_vendor_page.dart';
 export 'package:cybearjinni/presentation/pages/add_vendors_pages/add_ewelink_vendor/add_ewelink_vendor_page.dart';
 export 'package:cybearjinni/presentation/pages/add_vendors_pages/add_lifx_vendor/add_lifx_vendor_page.dart';
-export 'package:cybearjinni/presentation/pages/change_room_for_devices/change_room_for_devices_page.dart';
+export 'package:cybearjinni/presentation/pages/change_room_for_devices_page.dart';
+export 'package:cybearjinni/presentation/pages/choose_automation_type_to_add_page.dart';
 export 'package:cybearjinni/presentation/pages/connect_to_hub/connect_to_hub_page.dart';
-export 'package:cybearjinni/presentation/pages/connect_to_hub_more/connect_to_hub_more_page.dart';
+export 'package:cybearjinni/presentation/pages/connect_to_hub_more_page.dart';
 export 'package:cybearjinni/presentation/pages/entities_in_network_page.dart';
 export 'package:cybearjinni/presentation/pages/entities_in_room_page.dart';
 export 'package:cybearjinni/presentation/pages/home_page/home_page.dart';
 export 'package:cybearjinni/presentation/pages/introduction_screen/introduction_screen_page.dart';
-export 'package:cybearjinni/presentation/pages/plus_button/plus_button.dart';
-export 'package:cybearjinni/presentation/pages/remote_pipes/remote_pipes_page.dart';
+export 'package:cybearjinni/presentation/pages/pick_repeat_time_page.dart';
+export 'package:cybearjinni/presentation/pages/plus_button.dart';
+export 'package:cybearjinni/presentation/pages/remote_pipes_page.dart';
 export 'package:cybearjinni/presentation/pages/scenes/scenes_page.dart';
 export 'package:cybearjinni/presentation/pages/smart_camera_container_page.dart';
-export 'package:cybearjinni/presentation/pages/software_info/software_info_page.dart';
+export 'package:cybearjinni/presentation/pages/software_info_page.dart';
 export 'package:cybearjinni/presentation/pages/splash_page.dart';
 export 'package:cybearjinni/presentation/pages/video_stream_output_container_page.dart';
